requirements_kafka.txt
# requirements.txt
# Biblioteki potrzebne do projektu Altseason z Apache Kafka

# Kafka
kafka-python==2.0.2

# Exchanges and API
ccxt==4.1.77
requests==2.31.0

# Data processing
pandas==2.1.4
numpy==1.24.3

# Visualization
matplotlib==3.8.2
plotly==5.17.0

# Scientific computing
scipy==1.11.4

# Optional: for advanced analytics
scikit-learn==1.3.2

# Optional: for better JSON handling
ujson==5.8.0

# Optional: for better datetime handling
python-dateutil==2.8.2

# Optional: for configuration management
python-dotenv==1.0.0

# Development tools (optional)
jupyter==1.0.0
jupyterlab==4.0.9
